Remove some unit tests that are not relevant for Sobek and move and fix those that are relevant
