+++ sobek-service/src/test/java/org/rutebanken/sobek/auth/DefaultAuthorizationServiceTest.java
@@ -1,25 +1,29 @@
 package org.rutebanken.sobek.auth;
 
-import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.rutebanken.helper.organisation.RoleAssignment;
+import org.springframework.boot.test.context.SpringBootTest;
 
 import java.util.List;
 
-class DefaultAuthorizationServiceTest {
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+@SpringBootTest
+public class DefaultAuthorizationServiceTest {
 
     @Test
     void canEditAllEntities() {
         List<RoleAssignment> roleAssignments = RoleAssignmentListBuilder.builder().withAccessAllAreas().build();
         DefaultAuthorizationService defaultAuthorizationService = new DefaultAuthorizationService(null,false, null);
-        Assertions.assertTrue(defaultAuthorizationService.verifyCanEditAllEntities(roleAssignments));
+        assertTrue(defaultAuthorizationService.verifyCanEditAllEntities(roleAssignments));
     }
 
     @Test
     void canEditAllEntitiesMissingRoleAssignment() {
         List<RoleAssignment> roleAssignments = RoleAssignmentListBuilder.builder().build();
         DefaultAuthorizationService defaultAuthorizationService = new DefaultAuthorizationService(null,false, null);
-        Assertions.assertFalse(defaultAuthorizationService.verifyCanEditAllEntities(roleAssignments));
+        assertFalse(defaultAuthorizationService.verifyCanEditAllEntities(roleAssignments));
 
     }
 }
