+++ sobek-service/src/test/java/org/rutebanken/sobek/model/VersionOfObjectRefStructureTest.java
@@ -15,11 +15,13 @@
 
 package org.rutebanken.sobek.model;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 
+@SpringBootTest
 public class VersionOfObjectRefStructureTest {
     @Test
     public void equals() throws Exception {
