+++ sobek-service/src/test/resources/application.properties
@@ -1,5 +1,5 @@
 spring.datasource.driver-class-name=org.h2.Driver
-spring.datasource.url=jdbc:h2:mem:testdb;NON_KEYWORDS=VALUE;MODE=MySQL
+spring.datasource.url=jdbc:h2:mem:testdb;NON_KEYWORDS=KEY,VALUE;MODE=MySQL
 spring.datasource.username=sa
 spring.datasource.password=
