--- src/test/resources/application.properties
@@ -1,90 +1,0 @@
-#
-# Licensed under the EUPL, Version 1.2 or – as soon they will be approved by
-# the European Commission - subsequent versions of the EUPL (the "Licence");
-# You may not use this work except in compliance with the Licence.
-# You may obtain a copy of the Licence at:
-#
-#   https://joinup.ec.europa.eu/software/page/eupl
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the Licence is distributed on an "AS IS" basis,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the Licence for the specific language governing permissions and
-# limitations under the Licence.
-#
-
-logging.level.org.hibernate.tool.hbm2ddl=INFO
-logging.level.org.hibernate.SQL=WARN
-logging.level.org.hibernate.type=WARN
-logging.level.org.springframework=WARN
-logging.level.org.rutebanken.sobek=INFO
-logging.level.ma.glasnost.orika=WARN
-
-logging.level.org.rutebanken.sobek.importer=DEBUG
-
-logging.level.org.rutebanken.sobek.service.StopPlaceFromQuaysCorrelationService=WARN
-logging.level.org.rutebanken.sobek.importer.finder.StopPlaceFromOriginalIdFinder=TRACE
-
-spring.jpa.properties.hibernate.hbm2ddl.import_files_sql_extractor=org.hibernate.tool.hbm2ddl.MultipleLinesSqlCommandExtractor
-spring.jpa.properties.hibernate.ddl-auto=create-drop
-
-spring.jpa.database=POSTGRESQL
-
-spring.database.driverClassName=org.postgresql.Driver
-
-spring.datasource.url=jdbc:tc:postgis:13-3.3:///sobek
-
-spring.jpa.properties.jakarta.persistence.query.timeout=15000
-
-
-spring.jpa.hibernate.ddl-auto=update
-spring.jpa.open-in-view=false
-spring.jpa.show-sql=false
-
-spring.datasource.hikari.maximumPoolSize=30
-spring.datasource.username=sobek
-
-spring.flyway.enabled=true
-spring.flyway.table=schema_version
-
-blobstore.gcs.blob.path=exports
-blobstore.gcs.bucket.name=sobek-test
-
-blobstore.gcs.project.id=carbon-1287
-
-server.port=1888
-
-spring.profiles.active=test
-
-rutebanken.kubernetes.enabled=false
-
-#OAuth2 Resource Server
-sobek.oauth2.resourceserver.auth0.entur.internal.jwt.issuer-uri=http://notInUse
-sobek.oauth2.resourceserver.auth0.entur.internal.jwt.audience=http://notInUse
-
-netex.import.enabled.types=MERGE,INITIAL,ID_MATCH,MATCH
-netexXmlReferenceValidator.throwOnValidationError=true
-asyncNetexExport.validateAgainstSchema=true
-
-async.export.path=/tmp
-
-#For tariff zone import test
-stopPlaceRefUpdaterService.enableLegacyUpdater=true
-
-# pubsub
-spring.cloud.gcp.pubsub.project-id=test
-# Path to credential for accessing the Google Cloud project, using the Spring resource syntax (prefixed with file:)
-#spring.cloud.gcp.pubsub.credentials.location=file:/path/to/json/file
-# if this property is present then the client library connects to the PubSub emulator instead of Google Cloud
-spring.cloud.gcp.pubsub.emulatorHost=localhost:8089
-# number of parallel pullers for a given Subscriber. Constrained by the number of threads in the executor.
-spring.cloud.gcp.pubsub.subscriber.parallel-pull-count=1
-# number of executor threads (shared by all subscribers)
-spring.cloud.gcp.pubsub.subscriber.executor-threads=30
-# the client library will automatically extend the ack deadline up to this limit.
-spring.cloud.gcp.pubsub.subscriber.max-ack-extension-period=36000
-# number of threads performing ack deadline extension  (shared by all subscribers). The default value is 5
-#entur.pubsub.subscriber.system-threads=5
-# path to Google PubSub emulator
-# default path if the property is not set (retrieved as a Maven dependency):
-#entur.pubsub.emulator.path=target/pubsub-emulator/pubsub-emulator-0.1-SNAPSHOT.jar
