--- src/test/resources/db/migration/V0__setup.sql
@@ -1,2 +1,0 @@
-CREATE USER sobek with PASSWORD 'sobek';
-CREATE USER postgres;
