--- src/test/resources/org/rutebanken/tiamat/rest/netex/publicationdelivery/parking.xml
@@ -1,79 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<PublicationDelivery xmlns="http://www.netex.org.uk/netex">
-    <PublicationTimestamp>2025-04-22T10:58:08.484</PublicationTimestamp>
-    <ParticipantRef>NSR</ParticipantRef>
-    <Description>Pakring to </Description>
-    <dataObjects>
-        <SiteFrame id="NSR:SiteFrame:380382157" version="1">
-            <FrameDefaults>
-                <DefaultLocale>
-                    <TimeZone>Europe/Oslo</TimeZone>
-                    <DefaultLanguage>nor</DefaultLanguage>
-                </DefaultLocale>
-            </FrameDefaults>
-            <parkings>
-                <Parking created="2017-06-19T19:13:17.15" changed="2024-12-09T13:29:20.88" modification="new" version="9" id="NSR:Parking:70">
-                    <keyList>
-                        <KeyValue>
-                            <Key>imported-id</Key>
-                            <Value>NSB:Parking:216bike</Value>
-                        </KeyValue>
-                        <KeyValue>
-                            <Key>CHANGED_BY</Key>
-                            <Value>johan</Value>
-                        </KeyValue>
-                    </keyList>
-                    <Name lang="nor">Gøteborg</Name>
-                    <Centroid>
-                        <Location>
-                            <Longitude>11.971662</Longitude>
-                            <Latitude>57.709531</Latitude>
-                        </Location>
-                    </Centroid>
-                    <Covered>outdoors</Covered>
-                    <ParentSiteRef ref="NSR:StopPlace:408"/>
-                    <ParkingVehicleTypes>pedalCycle</ParkingVehicleTypes>
-                    <ParkingPaymentProcess/>
-                    <parkingProperties>
-                        <ParkingProperties version="0" id="NSR:ParkingProperties:477">
-                            <ParkingUserTypes/>
-                            <spaces>
-                                <ParkingCapacity version="0" id="NSR:ParkingCapacity:945">
-                                    <ParkingUserType>allUsers</ParkingUserType>
-                                </ParkingCapacity>
-                            </spaces>
-                        </ParkingProperties>
-                    </parkingProperties>
-                </Parking>
-                <Parking created="2024-05-13T09:57:21Z" changed="2024-10-10T04:29:46Z" modification="new"
-                         version="3" id="FSR:Parking:1">
-                    <keyList>
-                        <KeyValue>
-                            <Key>9ba44e2b-0f11-ef11-9f89-000d3a699c93</Key>
-                            <Value>11111</Value>
-                        </KeyValue>
-                    </keyList>
-                    <ParentSiteRef ref="NSR:StopPlace:408"/>
-                    <ParkingVehicleTypes>car</ParkingVehicleTypes>
-                    <ParkingLayout>openSpace</ParkingLayout>
-                    <TotalCapacity>10</TotalCapacity>
-                    <ParkingPaymentProcess>payByPrepaidToken</ParkingPaymentProcess>
-                    <parkingProperties>
-                        <ParkingProperties version="0" id="FSR:ParkingProperties:4">
-                            <ParkingUserTypes></ParkingUserTypes>
-                            <spaces>
-                                <ParkingCapacity version="0" id="FSR:ParkingCapacity:5">
-                                    <ParkingUserType>allUsers</ParkingUserType>
-                                    <NumberOfSpaces>10</NumberOfSpaces>
-                                </ParkingCapacity>
-                                <ParkingCapacity version="0" id="FSR:ParkingCapacity:6">
-                                    <ParkingUserType>registeredDisabled</ParkingUserType>
-                                </ParkingCapacity>
-                            </spaces>
-                        </ParkingProperties>
-                    </parkingProperties>
-                </Parking>
-            </parkings>
-        </SiteFrame>
-    </dataObjects>
-</PublicationDelivery>
